requirements.txt
streamlit==1.27.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pyperclip==1.8.2
